update check permission

+++ app/Http/Controllers/AnswerController.php
@@ -29,14 +29,15 @@ class AnswerController extends Controller
      */
     public function store(AnswerRequest $request)
     {
-        $question = new Answer();
-        $question->fill($request->all());
-        $question->save();
+        if ($request->user()->can('answer-create')) {
 
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Thêm mới thành công!'
-        ], 200);
+            $question = new Answer();
+            $question->fill($request->all());
+            $question->save();
+
+            return statusResponse(200,"Thêm mới thành công!");
+        }
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 
     /**
@@ -44,26 +45,26 @@ class AnswerController extends Controller
      */
     public function update(AnswerRequest $request)
     {
-        $question = Answer::findOrFail($request->id);
-        $question->fill($request->all());
-        $question->save();
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Cập nhật thành công!'
-        ], 200);
+        if ($request->user()->can('answer-edit')) {
+
+            $question = Answer::findOrFail($request->id);
+            $question->fill($request->all());
+            $question->save();
+            return statusResponse(200,"Cập nhật thành công!");
+        }
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
-        $question = Answer::find($id);
-        Answer::destroy( $question->id);
-        
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Xóa câu hỏi thành công!'
-        ], 200);
+        if ($request->user()->can('answer-delete')) {
+            $question = Answer::find($id);
+            Answer::destroy( $question->id);
+            return statusResponse(200,"Xóa đáp án thành công!");
+        }
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 }
