+++ app/Http/Controllers/LessonController.php
@@ -74,32 +74,32 @@ class LessonController extends Controller
      */
     public function store(LessonRequest $request)
     {
-        $course = Course::where('slug', $request->course_slug)->first();
-        $lesson = new Lesson();
-        $lesson->fill($request->all());
-        $lesson->link = 'https://www.youtube.com/embed/'. basename($request->link);
-        $lesson->course_id = $course->id;
-        $lesson->save();
+        if ($request->user()->can('lesson-create')) {
 
-        if ($request->hasfile('file')) {
-            $username = Auth::user()->username;
+            $course = Course::where('slug', $request->course_slug)->first();
+            $lesson = new Lesson();
+            $lesson->fill($request->all());
+            $lesson->link = 'https://www.youtube.com/embed/'. basename($request->link);
+            $lesson->course_id = $course->id;
+            $lesson->save();
 
-            $listFile =  $this->imageService->fileUpload($request->file()['file'], 'public/docs/' . $username);
-            foreach ($listFile as $file) {
-                $fileUpload = new FileUpload();
-                $fileUpload->name = $file->name;
-                $fileUpload->type = $file->type;
-                $fileUpload->lesson_id = $lesson->id;
-                $fileUpload->file_path = "public/docs/" . $username . '/' . $fileUpload->name;
-                $fileUpload->name_table = "lessons";
-                $fileUpload->save();
+            if ($request->hasfile('file')) {
+                $username = Auth::user()->username;
+
+                $listFile =  $this->imageService->fileUpload($request->file()['file'], 'public/docs/' . $username);
+                foreach ($listFile as $file) {
+                    $fileUpload = new FileUpload();
+                    $fileUpload->name = $file->name;
+                    $fileUpload->type = $file->type;
+                    $fileUpload->lesson_id = $lesson->id;
+                    $fileUpload->file_path = "public/docs/" . $username . '/' . $fileUpload->name;
+                    $fileUpload->name_table = "lessons";
+                    $fileUpload->save();
+                }
             }
+            return statusResponse(200 ,"Thêm mới thành công!");
         }
-
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Thêm mới thành công!'
-        ], 200);
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 
     /**
@@ -131,33 +131,35 @@ class LessonController extends Controller
      */
     public function update(Request $request)
     {
-        $lesson = Lesson::findOrFail($request->id);
-        $lesson->fill($request->all());
-        $lesson->save();
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Cập nhật thành công!'
-        ], 200);
+        if ($request->user()->can('lesson-update')) {
+
+            $lesson = Lesson::findOrFail($request->id);
+            $lesson->fill($request->all());
+            $lesson->save();
+            return statusResponse(200 ,"Cập nhật thành công!");
+        }
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $slug)
+    public function destroy(Request $request, string $slug)
     {
-        $lesson = Lesson::where('slug', $slug)->first();
-        $file_uploads = $lesson->fileUploads;
+        if ($request->user()->can('lesson-delete')) {
 
-        if ($file_uploads) {
-            $this->imageService->removeFileInStorage($file_uploads);
+            $lesson = Lesson::where('slug', $slug)->first();
+            $file_uploads = $lesson->fileUploads;
+
+            if ($file_uploads) {
+                $this->imageService->removeFileInStorage($file_uploads);
+            }
+
+            Lesson::destroy($lesson->id);
+
+            return statusResponse(200 ,"Xóa bài học thành công!");
         }
-
-        Lesson::destroy($lesson->id);
-
-        return response()->json([
-            'statusCode' => 200,
-            'message' => 'Xóa bài học thành công!'
-        ], 200);
+        return statusResponse(401,"Bạn không có quyền truy cập");
     }
 
     public function fileUpload(Request $request)
